Application Page
I added the application form, which right now the user will be taken to after they login. This page includes the fields they will have to fill out on each screen with previous and next button functionality. I also added a progress bar on the screens. Finally, it includes a submit button and submission page to confirm the application is submitted.